Initial stab at setting GPIO speed.
Created initial implementation of setting port speed. Trying out enums as an alternative to making functions for every little thing and sizing them to the size of the register even if they operate on a subset to prevent unnecessary casting. (Also in experience aligning things to registers generally performs better).